Move BranchCache into store
It should not be exported at top level.
